refactor(test): co-locate shared test-support and consolidate YGOProRoom
Move the shared Object Mothers and Mock classes from tests/modules/shared/
to src/test-support/ (git mv preserves history), the canonical home per
docs/testing.md. Add a @test-support/* path alias (in both tsconfig.json and
tsconfig.eslint.json) and update the 4 legacy importers.

Exclude **/*.test.ts and src/test-support/** from the production tsconfig so
tsc no longer emits test code into dist/ — a pre-existing leak now that test
support lives under src/. Jest is unaffected (ts-jest compiles independently).

Consolidate the 3 inline YGOProRoom factories (YGOProRoomFinalizing,
YGOProRoomWindbotFlags, FinalizeYGOProRoom) onto YGOProRoomMother, extended
to accept overrides. YGOProRoom is no longer built two different ways.

504 tests green, build clean, no test artifacts in dist after rebuild.
